+++ src/main/java/pf/yozzman/risk/phase/PhaseAttaque.java
@@ -25,11 +25,22 @@ public class PhaseAttaque {
     public void executer() {
         String choix = lecteur.lireLigne("Voulez-vous attaquer ? (o/n) ");
         while (choix.equalsIgnoreCase("o")) {
+            //String from = lecteur.lireLigne("Pays attaquant (nom exact) : ");
+            //String to = lecteur.lireLigne("Pays défenseur (nom exact) : ");
+            
             String from = lecteur.lireLigne("Pays attaquant (nom exact) : ");
-            String to = lecteur.lireLigne("Pays défenseur (nom exact) : ");
-
             Pays pFrom = carte.getPaysByName(from);
+
+            // Si le pays attaquant est valide et vous appartient, afficher seulement les voisins ennemis
+            if (pFrom != null && pFrom.getProprietaire() == joueur) {
+                carte.afficherCarteVoisin(pFrom);
+            }
+
+            String to = lecteur.lireLigne("Pays défenseur (nom exact) : ");
             Pays pTo = carte.getPaysByName(to);
+
+            pFrom = carte.getPaysByName(from);
+            pTo = carte.getPaysByName(to);
 
             if (!verifierConditions(pFrom, pTo)) {
                 choix = lecteur.lireLigne("Autre attaque ? (o/n) ");
