+++ src/main/java/pf/yozzman/risk/util/ConsoleWriter.java
